Fix for crash in hf_generator function in lpp_trans
Number of envelopes is becoming zero because of erroneous input
stream.Inside SBR start band and stop band are calculated based
on number of envelope's.

In this case end position is becoming zero, which in turn makes
start bands as negative. In sbr processing buffer is accessed
from start to stop band. This is causing OOB read access

Bug:114744962
Test: poc
Change-Id: Ib51fc464d0afc2f5a68d860fcde1b8961fd69d40