Fix for Parametric stereo stream playback issue
The stream used for this issue had parametric stereo switched off
for first few frames and decoder is configured to give mono output.
After few frames even though codec has detected PS content in the stream
channel configuration is not getting updated which caused breaks in
audio.

As a fix we are checking future possibility on PS content in the stream
based on AOT and using that information to update output channel count.

Bug:114329895
Test: vendor
Change-Id: I0faa47ff234b8651e2c42f9e4423fe9bf1c3908c