Fix for testbench crash due to invalid file handle
Added checks to handle cases when input, output or
meta files are not passed in arguments of xaacdec.
Removed process command mode support and related
code from testbench.
Removed interface file read support and moved the
condition for drc to be on, if effect type or target
loudness is present in the command line arguments.

Bug:124627497
Test: poc
Change-Id: I97e216f5bc6ba6c52dcdc6438ee4152f0cb1682c